Segemation of Tasklist into a new class, away from Bird
